requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
filelock==3.16.1
Flask==3.0.3
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.15
langchain-mistralai==0.2.1
langsmith==0.1.139
MarkupSafe==3.0.2
numpy==2.1.3
orjson==3.10.11
packaging==24.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
scikit-learn==1.5.2
scipy==1.14.1
sniffio==1.3.1
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.20.1
tqdm==4.66.6
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.1
